Convert UniswapInfoActivity to fragment
