--- app/src/main/java/io/horizontalsystems/bankwallet/modules/swap/view/UniswapInfoActivity.kt
@@ -1,49 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.swap.view
-
-import android.app.Activity
-import android.content.Intent
-import android.net.Uri
-import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
-import io.horizontalsystems.bankwallet.R
-import io.horizontalsystems.bankwallet.core.BaseActivity
-import kotlinx.android.synthetic.main.activity_swap_info.*
-
-class UniswapInfoActivity : BaseActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_swap_info)
-
-        setSupportActionBar(toolbar)
-
-        buttonUniswap.setOnClickListener {
-            val intent = Intent(Intent.ACTION_VIEW)
-            intent.data = Uri.parse("https://uniswap.org/")
-            startActivity(intent)
-        }
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.select_coin_menu, menu)
-        return true
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item.itemId) {
-            R.id.menuClose -> {
-                finish()
-                return true
-            }
-        }
-        return super.onOptionsItemSelected(item)
-    }
-
-    companion object {
-        fun start(activity: Activity) {
-            val intent = Intent(activity, UniswapInfoActivity::class.java)
-            activity.startActivity(intent)
-        }
-    }
-}
