--- app/src/main/res/menu/select_coin_menu.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/menuClose"
-        android:title="@string/Button_Close"
-        app:showAsAction="always" />
-</menu>
